refactor: move console, json_output, projection_diff into output/ package
Bundles all output-related modules under src/poly/output/:
- console.py → output/console.py (Rich formatting)
- output.py → output/json_output.py (JSON helpers)
- projection_diff.py → output/projection_diff.py (proto diff logic)

Re-exports everything from output/__init__.py for clean imports.
